simple modify

+++ infiniswap_bd/dashboard.h
@@ -11,17 +11,18 @@
 #define EXCEPTION_RATIO 0.05
 #define MAX_RW_SIZE 20000
 
-struct bd_info{
+struct bd_info
+{
     unsigned long long avg_read_latency;
     unsigned long long avg_write_latency;
     unsigned long long high_ex_read_latency;
     unsigned long long low_ex_read_latency;
     unsigned long long high_ex_write_latency;
     unsigned long long low_ex_write_latency;
-    unsigned long long high_read_latency[EXCEPTION_RATIO * MAX_RW_SIZE];
-    unsigned long long low_read_latency[EXCEPTION_RATIO * MAX_RW_SIZE];
-    unsigned long long high_write_latency[EXCEPTION_RATIO * MAX_RW_SIZE];
-    unsigned long long low_write_latency[EXCEPTION_RATIO * MAX_RW_SIZE];
+    unsigned long long high_read_latency[(int)(EXCEPTION_RATIO * MAX_RW_SIZE)];
+    unsigned long long low_read_latency[(int)(EXCEPTION_RATIO * MAX_RW_SIZE)];
+    unsigned long long high_write_latency[(int)(EXCEPTION_RATIO * MAX_RW_SIZE)];
+    unsigned long long low_write_latency[(int)(EXCEPTION_RATIO * MAX_RW_SIZE)];
     unsigned read_num;
     unsigned write_num;
     unsigned request_num;
@@ -36,7 +37,7 @@ void add_latency(unsigned long long latency, int write);
 
 void clear_info(void);
 
-int write_to_file(void); 
+int write_to_file(void);
 
 void write_info(void);
